pyproject.toml
[tool.poetry]
name = "screengenie"
version = "0.1.0"
description = "context aware ai assistant"
authors = ["gayan ranasinghe <gkranasinghe3@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ollama = "^0.4.8"
pyyaml = "^6.0.2"
pytesseract = "^0.3.13"
pillow = "^11.2.1"
watchdog = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
